Rework the connection pool handling
The previous implementation was quite buggy or at best not working as expected.
Now, a connection pool is created on application start up and bound to the YuDatabase class which uses it to get and return connections from/to the pool.
Additionally, on checking out a connection, the connection itself is checked whether it is still valid. This is done by a simple ping().
If it fails, the connection is re-established before it is returned to the caller.